+++ automatrix/__init__.py
@@ -1,0 +1,7 @@
+"""A module for performing matrix operations and showing the work for them."""
+from automatrix.commands import *
+from automatrix.dispatcher import Dispatcher
+from automatrix.latex import LatexInterface
+from automatrix.matrix import Matrix
+
+__all__ = ["Dispatcher", "LatexInterface", "Matrix"]
