+++ automatrix/commands/__init__.py
@@ -1,0 +1,12 @@
+"""A module for the automatrix commands."""
+import automatrix.commands.inverse2x2 as inverse2x2
+import automatrix.commands.inverserref as inverserref
+import automatrix.commands.matrixmultiply as matrixmultiply
+import automatrix.commands.patterns as patterns
+
+__all__ = [
+    "inverse2x2",
+    "inverserref",
+    "matrixmultiply",
+    "patterns",
+]
